backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
pandas
numpy
scipy
openpyxl
python-comtrade
